Add resolve_slicing function for time and log slicing validation
