+++ src/drtsans/mono/gpsans/api.py
@@ -23,9 +23,10 @@ import matplotlib.pyplot as plt
 from drtsans import getWedgeSelection, subtract_background, NoDataProcessedError
 from drtsans.beam_finder import center_detector, fbc_options_json, find_beam_center
 from drtsans.dataobjects import save_i1d
+from drtsans.eventslice import resolve_slicing
 from drtsans.instruments import extract_run_number, instrument_filesystem_name
 from drtsans.iq import bin_all
-from drtsans.load import move_instrument, resolve_slicing
+from drtsans.load import move_instrument
 from drtsans.polarization import PolarizationLevel
 from drtsans.mask_utils import apply_mask, load_mask
 from drtsans.mono import meta_data
@@ -255,25 +256,26 @@ def load_all_files(
         if not registered_workspace(ws_name):
             filename = abspath(sample.strip(), instrument=instrument_name, ipts=ipts, directory=path)
             logger.notice(f"Loading filename {filename} to slice")
+            timesliceinterval = timesliceperiod = logslicename = logsliceinterval = None
             if timeslice:
                 timesliceinterval = reduction_config["timeSliceInterval"]
                 timesliceperiod = reduction_config["timeSlicePeriod"]
-                logslicename = logsliceinterval = None
             elif logslice:
-                timesliceinterval, timesliceperiod = None, None
                 logslicename = reduction_config["logSliceName"]
                 logsliceinterval = reduction_config["logSliceInterval"]
+            elif polarized:
+                pass  # polarization metadata stored in reduction_config['polarization']
 
             load_and_split(
-                filename,
+                filename,  # file path to the sample run
+                sample_to_si_name=SAMPLE_SI_META_NAME,
+                si_nominal_distance=SI_WINDOW_NOMINAL_DISTANCE_METER,
                 output_workspace=ws_name,
                 time_interval=timesliceinterval,
                 time_offset=reduction_config["timeSliceOffset"],
                 time_period=timesliceperiod,
                 log_name=logslicename,
                 log_value_interval=logsliceinterval,
-                sample_to_si_name=SAMPLE_SI_META_NAME,
-                si_nominal_distance=SI_WINDOW_NOMINAL_DISTANCE_METER,
                 sample_to_si_value=swd_value_dict[meta_data.SAMPLE],
                 sample_detector_distance_value=sdd_value_dict[meta_data.SAMPLE],
                 reduction_config=reduction_config,
